Upgrade to Devel::PPPort 3.01.
p4raw-id: //depot/perl@23234

+++ ext/Devel/PPPort/t/ppphtest.t
@@ -57,8 +57,8 @@ if ($perl =~ m!^\./!) {
 }
 
 END {
-  chdir("..") if !-d $tmp && -d "../$tmp";
-  rmtree($tmp);
+  chdir('..') if !-d $tmp && -d "../$tmp";
+  rmtree($tmp) if -d $tmp;
 }
 
 ok(&Devel::PPPort::WriteFile("ppport.h"));
